pyproject.toml
[tool.poetry]
name = "data"
packages = [
    { include = "src", from = "."}
]
version = "0.1.0"
description = ""
authors = ["patrickbott <gorays999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.13"
pandas = "^2.2.3"
yfinance = "^0.2.48"
databento = "^0.44.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
psycopg2 = "^2.9.10"
sqlalchemy = "1.4.36"
apache-airflow = "^2.10.4"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pyarrow = "^18.1.0"
pydantic = "^2.10.4"
tzdata = "^2024.2"
prometheus-fastapi-instrumentator = "^7.0.0"
slowapi = "^0.1.9"
mkdocstrings-python = "^1.13.0"
streamlit = "^1.49.1"
psycopg2-binary = "^2.9.10"
prometheus-client = "^0.22.1"


[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = "^0.27.0"
mkdocs-material = "^9.5.49"
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
